# openapi.json
{
  "components": {
    "schemas": {
...
      "CorreccionEntrada": {
        "additionalProperties": false,
        "properties": {
          "correcciones": {
            "additionalProperties": {
              "type": "string"
            },
            "maxProperties": 12,
            "minProperties": 1,
            "title": "Correcciones",
            "type": "object"
          }
        },
        "required": [
          "correcciones"
        ],
        "title": "CorreccionEntrada",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Respuesta_RegistroSalida_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RegistroSalida"
          },
          "meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MetaDePagina"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "default": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Respuesta[RegistroSalida]",
        "type": "object"
      },
...
      "Respuesta_list_RegistroSalida__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/RegistroSalida"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MetaDePagina"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "default": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Respuesta[list[RegistroSalida]]",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Ingesta de adjuntos de correo y extraccion de datos tributarios. Todas las rutas requieren un token OIDC salvo las sondas de salud.",
    "title": "Automatizacion de Correos API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/records": {
      "get": {
        "description": "Listado principal, paginado por cursor y con filtros combinables.",
        "operationId": "registros_listar_registros",
        "parameters": [
          {
            "in": "query",
            "name": "trabajo_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trabajo Id"
            }
          },
          {
            "in": "query",
            "name": "ruc",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 11,
                  "pattern": "^\\d{11}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ruc"
            }
          },
          {
            "in": "query",
            "name": "ruc_inquilino",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 11,
                  "pattern": "^\\d{11}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ruc Inquilino"
            }
          },
          {
            "in": "query",
            "name": "periodo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 6,
                  "pattern": "^\\d{6}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Periodo"
            }
          },
          {
            "in": "query",
            "name": "periodo_desde",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 6,
                  "pattern": "^\\d{6}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Periodo Desde"
            }
          },
          {
            "in": "query",
            "name": "periodo_hasta",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 6,
                  "pattern": "^\\d{6}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Periodo Hasta"
            }
          },
          {
            "in": "query",
            "name": "fecha_desde",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fecha Desde"
            }
          },
          {
            "in": "query",
            "name": "fecha_hasta",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fecha Hasta"
            }
          },
          {
            "in": "query",
            "name": "solo_aprobados",
            "required": false,
            "schema": {
              "default": false,
              "title": "Solo Aprobados",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limite",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limite",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Respuesta_list_RegistroSalida__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listar Registros",
        "tags": [
          "registros"
        ]
      }
    },
    "/api/v1/records/{registro_id}": {
      "get": {
        "operationId": "registros_obtener_registro",
        "parameters": [
          {
            "in": "path",
            "name": "registro_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Registro Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Respuesta_RegistroSalida_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Obtener Registro",
        "tags": [
          "registros"
        ]
      },
      "patch": {
        "description": "Aplica las correcciones de una persona y aprueba el registro.\n\nLos valores llegan como texto y pasan por los mismos objetos de\nvalor que la extraccion automatica: una persona tambien teclea mal\nun RUC, y aceptarlo sin verificar dejaria entrar justo el error\nque el pipeline existe para evitar.",
        "operationId": "registros_corregir_registro",
        "parameters": [
          {
            "in": "path",
            "name": "registro_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Registro Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorreccionEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Respuesta_RegistroSalida_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Corregir Registro",
        "tags": [
          "registros"
        ]
      }
    },
    "/api/v1/records/{registro_id}/approve": {
      "post": {
        "description": "El revisor confirma que lo extraido es correcto.",
        "operationId": "registros_aprobar_registro",
        "parameters": [
          {
            "in": "path",
            "name": "registro_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Registro Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Respuesta_RegistroSalida_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Aprobar Registro",
        "tags": [
          "registros"
        ]
      }
    },
    "/api/v1/records/{registro_id}/document": {
      "get": {
        "description": "URL prefirmada de vida corta para ver el adjunto original.\n\nSe verifica primero que el registro es del tenant (y existe) antes de\nfirmar nada: asi el identificador del adjunto no sirve para sondear\ndocumentos ajenos.",
        "operationId": "registros_ver_documento",
        "parameters": [
          {
            "in": "path",
            "name": "registro_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Registro Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Respuesta_DocumentoDeAdjuntoSalida_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ver Documento",
        "tags": [
          "registros"
        ]
      }
    },
    "/api/v1/records/{registro_id}/reject": {
      "post": {
        "description": "Marca el registro como no utilizable: documento ilegible o ajeno.",
        "operationId": "registros_rechazar_registro",
        "parameters": [
          {
            "in": "path",
            "name": "registro_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Registro Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Respuesta_RegistroSalida_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rechazar Registro",
        "tags": [
          "registros"
        ]
      }
    },
    "/api/v1/reports/exports": {
      "post": {
        "description": "Encola la generacion del reporte y devuelve su identificador.",
        "operationId": "registros_solicitar_exportacion",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitarExportacionEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Respuesta_ExportacionSalida_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Solicitar Exportacion",
        "tags": [
          "registros"
        ]
      }
    },
    "/api/v1/reports/exports/{exportacion_id}": {
      "get": {
        "description": "Estado de la exportacion y, cuando esta lista, su URL de descarga.\n\nLa URL vive cinco minutos: lleva los datos tributarios del cliente\ny suele acabar pegada en un chat.",
        "operationId": "registros_consultar_exportacion",
        "parameters": [
          {
            "in": "path",
            "name": "exportacion_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Exportacion Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Respuesta_ExportacionSalida_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Consultar Exportacion",
        "tags": [
          "registros"
        ]
      }
    },
    "/api/v1/review": {
      "get": {
        "description": "Registros que el pipeline no pudo dar por buenos.",
        "operationId": "registros_cola_de_revision",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limite",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limite",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Respuesta_list_RegistroSalida__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cola De Revision",
        "tags": [
          "registros"
        ]
      }
    },
...
  }
}